requirements.txt
flask==2.0.1
pyyaml==6.0
